eval/requirements.txt
ragas>=0.2.0
langchain-google-genai>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
httpx>=0.27.0
pandas>=2.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
